Update the cicd files 3.9
